requirements.txt
Flask==3.0.2
requests==2.31.0
fastapi==0.115.8
uvicorn==0.30.6
elevenlabs==1.50.2
google-generativeai==0.8.4
